refactor: extract EngineInfo class to common
EngineInfo was a subclass of:
- IsochronesResponseInfo
- MatrixResponseInfo
- RouteResponseInfo

It is now a common class for all responses
