+++ ors-api/src/main/java/org/heigit/ors/api/responses/routing/RouteResponseInfo.java
@@ -23,10 +23,10 @@ import io.swagger.v3.oas.annotations.media.Schema;
 import org.heigit.ors.api.EndpointsProperties;
 import org.heigit.ors.api.SystemMessageProperties;
 import org.heigit.ors.api.requests.routing.RouteRequest;
+import org.heigit.ors.api.responses.common.engineinfo.EngineInfo;
 import org.heigit.ors.api.util.AppInfo;
 import org.heigit.ors.api.util.SystemMessage;
 import org.heigit.ors.config.AppConfig;
-import org.json.JSONObject;
 
 @Schema(description = "Information about the request")
 @JsonInclude(JsonInclude.Include.NON_DEFAULT)
@@ -85,34 +85,4 @@ public class RouteResponseInfo {
         engineInfo.setGraphDate(graphDate);
     }
 
-    @Schema(description = "Information about the version of the openrouteservice that was used to generate the route")
-    private static class EngineInfo {
-        @Schema(description = "The backend version of the openrouteservice that was queried", example = "5.0")
-        @JsonProperty("version")
-        private final String version;
-        @Schema(description = "The date that the service was last updated", example = "2019-02-07T14:28:11Z")
-        @JsonProperty("build_date")
-        private final String buildDate;
-        @Schema(description = "The date that the graph data was last updated", example = "2019-02-07T14:28:11Z")
-        @JsonProperty("graph_date")
-        private String graphDate;
-
-        public EngineInfo(JSONObject infoIn) {
-            version = infoIn.getString("version");
-            buildDate = infoIn.getString("build_date");
-            graphDate = "0000-00-00T00:00:00Z";
-        }
-
-        public void setGraphDate(String graphDate) {
-            this.graphDate = graphDate;
-        }
-
-        public String getVersion() {
-            return version;
-        }
-
-        public String getBuildDate() {
-            return buildDate;
-        }
-    }
 }
